requirements.txt
autopep8==1.2
EasyProcess==0.1.9
Flask==0.10.1
flask-heroku==0.1.9
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.1
pep8==1.6.2
requests==2.7.0
requests-oauthlib==0.5.0
selenium==2.47.1
six==1.9.0
tweepy==3.3.0
Werkzeug==0.10.4
wheel==0.24.0
